Fix build errors with Visual Studio 2015 Update 1
Replace operator synonyms. We need to specify /Za option to use
operator synonyms with Visual Studio. But the /Za option causes
the compile error C2970 of passing a object with internal linkage
to a template non-type argument. I could not figure out how to
fix C2970, so I choose not to use the /Za option.