Fix a dangling-pointer reference
The default move constructor or assignment operator copies the pointer.
When a prepared_statement_handle_t object is moved, the sqlite_statement
of the moved-from object is finalized in its destructor while the moved-to
object's sqlite_statement still points to the finalized object. The
moved-to object ends up with a dangling pointer.

We need to explictly set sqlite_statement to nullptr to avoid the
destruction in the moved-from object's destructor.